CHROMIUM: Event log support in the chromeos_arm platform driver.
An in memory buffer is set up covering a range of addresses specified in the
device tree, and when a panic happens, an event log event is generated and
stored in that buffer. When the system reboots as a result of the panic, the
firmware discovers the event and records it in the event log in flash.

I've also added a remove function which may or may not be desirable.

BUG=chrome-os-partner:16132
TEST=Built and booted on spring, triggered a panic using /proc/sysrq-trigger,
and after the reboot, used mosys eventlog to verify that a panic event had
been recorded.

Change-Id: I2a921ae96fa497de93d40d10b64bf8d3d159c50a
Signed-off-by: Gabe Black <gabeblack@google.com>
(cherry picked from commit e950860a45b7a2c6e0fa5488ddfae78db33a9f46)
Reviewed-on: https://chromium-review.googlesource.com/66912
Reviewed-by: Olof Johansson <olofj@chromium.org>