UPSTREAM: ipv6: udp packets following an UFO enqueued packet need also be handled by UFO
In the following scenario the socket is corked:
If the first UDP packet is larger then the mtu we try to append it to the
write queue via ip6_ufo_append_data. A following packet, which is smaller
than the mtu would be appended to the already queued up gso-skb via
plain ip6_append_data. This causes random memory corruptions.

In ip6_ufo_append_data we also have to be careful to not queue up the
same skb multiple times. So setup the gso frame only when no first skb
is available.

This also fixes a shortcoming where we add the current packet's length to
cork->length but return early because of a packet > mtu with dontfrag set
(instead of sutracting it again).

Found with trinity.

Cc: YOSHIFUJI Hideaki <yoshfuji@linux-ipv6.org>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>
Reported-by: Dmitry Vyukov <dvyukov@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

BUG=chromium:426987
TEST=kernel compiles and boots

(cherry picked from commit 2811ebac2521ceac84f2bdae402455baa6a7fb47)
Signed-off-by: Stephen Barber <smbarber@chromium.org>

Change-Id: If4c294c52a78e22a81d2839793ff8b3653786187
Reviewed-on: https://chromium-review.googlesource.com/234458
Trybot-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>