CHERRY-PICK: MALI: Return 1 from pm_callback_power_on
The documentation for power_on_callback states:
	"@return 1 if the GPU state may have been lost, 0 otherwise."

Kbase uses this value to avoid resetting some state when the return
value is 0. If we do not reset the state and the GPU has been powered
down we may see hangs, as hardware chicken bits become unset.

This patch ensures correctness by always resetting the GPU state.
It may be possible to later optimize this to only reset the state as
required. However this optimization has been known to cause subtle
problems due to race conditions that may negate the slight benefit of
the optimization.

BUG=chrome-os-partner:22230, chrome-os-partner:23612
TEST=Sanity testing only: ensure basic functionality and browsing remains
unaffected. Testing to see if this resolves the BUG can only be done
once the change has landed.

Change-Id: Ib3639f75343e5cb9285681eb0cbb2f2511b1b60f
Signed-off-by: David Garbett <david.garbett@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/175123
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>