CHROMIUM: Add a new iTCO vendor support option for ChromeOS to disable TCO SMI
This new vendor support is selected by a kernel command line parameter on
x86/x86_64 Chrome OS. It disables TCO SMI when the watchdog timer is enabled.
On some existing devices, this workaround is needed because the TCO SMI handler
in coreboot clears the TIMEOUT bit of TCO1_STS register, preventing the machine
from rebooting when the timer expires. This workaround makes the watchdog work
on existing devices, but it is not necessary for devices with updated firmware.

BUG=chromium:321832
TEST=platform_HWwatchdog on mario,zgb,alex,parrot,stump,lumpy,butterfly

Change-Id: Iacb855476ca208d914bc61f9b20e03354e1377af
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184750
Reviewed-by: Aaron Durbin <adurbin@chromium.org>