BACKPORT: ASoC: Add macros for defining virtual enums
With the upcoming consolidation of normal MUXs and virtual MUXs we need to be
able to distinguish between enums with and without a backing register at the
enum level. Use the same approach as used for virtual mixer controls by setting
the reg field of the enum to SND_SOC_NOPM for enums without a backing register.
This patch adds a set of helper macros that can be used to define such enums.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit b948837a32cc2a510c2baf90111c8c5dde801a66)

Conflicts:
	include/sound/soc.h

Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

Change-Id: I4a1c3c7493445633ad97d77851d20103230d4907