CHROMIUM: ramoops: fix probe
The driver wasn't registered as most platforms drivers are -- it's just a one-shot
probe. So if the device is registered after the driver, probe won't happen. Fix that.

Also add a couple of helpful printks while I touch this driver; the current code has
moved in upstream/later kernels so this is just for our local kernel at the moment.

BUG=chromium:349122
TEST=run KernelErrorPaths

Change-Id: I79dd516781eb22172478f3501df310915e6162f8
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188797