BACKPORT: ASoC: Add a per component dai list
Now that every DAI has a component we can track the DAIs on a per component
basis. This simplifies the DAI lookup when we are only interested in DAIs of a
specific component and also makes it possible to have multiple components with
the same parent device and also register DAIs.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 1438c2f60ba955114cff3717f1a334878c7886a9)

Conflicts:
	sound/soc/soc-core.c

Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

Change-Id: Ie9c06f58b977b1a886d562be9122d49b30466da6