BACKPORT: net: Use netlink_ns_capable to verify the permisions of netlink messages
[ Upstream commit 90f62cf30a78721641e08737bda787552428061e ]

It is possible by passing a netlink socket to a more privileged
executable and then to fool that executable into writing to the socket
data that happens to be valid netlink message to do something that
privileged executable did not intend to do.

To keep this from happening replace bare capable and ns_capable calls
with netlink_capable, netlink_net_calls and netlink_ns_capable calls.
Which act the same as the previous calls except they verify that the
opener of the socket had the desired permissions as well.

Reported-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from v3.10.45 commit 1141a455802884d3bcbcf6b30e1d65d09cf286e1)
[cernekee: since ns_capable wasn't used back in v3.4, this patch just
 changes bare capable() into netlink_capable() so that the socket's
 original owner is checked.]

TEST=cbuildbot + manual boot
BUG=chromium:427008

Change-Id: I4b869cab7eb79fb5f12b9367b0b9ac80580dd306
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243623
Reviewed-by: Kees Cook <keescook@chromium.org>