UPSTREAM: ASoC: dapm: Simplify snd_soc_dapm_link_dai_widgets()
If we find a widget who's stream name matches the name of a DAI widget then
thats the one it should be connected to. Based on the widget id we can say in
which direction the path should be. No need to go back to the DAI and check the
stream names.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 0f9bd7b19467b9454a8e3729406f4cdcaaa32814)
Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

Change-Id: I618c47620d085e714491506cc88372b5faa3c94b