BACKPORT: dma: dw: add adsp register/unregister APIs for built in engine
The audio DSP in Haswell and Broadwell based platforms has a built in DesignWare
DMA engine. This engine is both used by the host and by the ADSP FW.

Allow the ADSP to register it's built in DMA engine so we can use the existing
DW DMA engine driver. The ADSP does not have a hclk, so register a dummy hclk
to allow probing of the DW driver.

Signed-off-by: Jie Yang <yang.jie@intel.com>
(cherry picked from commit 67146e68cbc3d23cc9e595f93f1c74c29a2abddc)
Signed-off-by: Kevin Strasser <kevin.strasser@intel.com>

Change-Id: Id7639dcbb99a53953de412d8891bdd4c510f5dcd