BACKPORT: drm/mipi-dsi: Add dual-channel support
This adds basic support for dual-channel support. It allows a pair of
devices to be linked in a master/slave relationship and provides helpers
to set this up. The master can optionally provide callbacks that are
executed when a slave is attached or removed.

Signed-off-by: Thierry Reding <treding@nvidia.com>

BUG=chrome-os-partner:30799
TEST=Tested on Ryu

(cherry picked from commit 7706176b0a40245c6ef919841c383bc6f6d716c8)
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Iab21bb8e64d05033a9fd839982ed46ff5d88b819
Reviewed-on: https://chromium-review.googlesource.com/216673
Reviewed-by: Benson Leung <bleung@chromium.org>