Add common kernel config options for virtio scsi support (needed for GCE).
This CL includes components that are needed for running Chromium OS in
Google Compute Engine and these components are explicitly included
in the kernel not as modules.

BUG=chromium:391105
TEST=Tested with changes in: cl:206583 in addition to uploading
the vm image to gce, creating an instance, and ssh'ing into it.
To use: USE=gce build_packages, build_image test, image_to_vm --test_image.

Change-Id: I4205dea25b6e690654d15996e60f8199938b4a38
Signed-off-by: Chris Sosa <sosa@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/211667
Reviewed-by: Olof Johansson <olofj@chromium.org>