CHROMIUM: ARM: tegra: dvfs: update set rate and add get_rate
Update set_rate to just return 0 if a clock is not subject to DVFS.
Also remove the warning. Implement a new get_rate() function which
returns the current clock rate as used for determining the required
voltage. This is used by is_clk_prepared() to determine if a clock is
used or not. We can't use clk_enabled() for this because
clk_disable_unused will disable unused clocks before unpreparing them,
so those clocks would stay prepared and contribute to the minimum
voltage requirement.

Corresponding 3.10 patch was reviewed on:
https://chromium-review.googlesource.com/179307

BUG=chrome-os-partner:31210
TEST=compile and boot

Change-Id: I8feb130bce20a2478781fdc30cb2e3605162ce5d
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Signed-off-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/214738
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217654
Tested-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>