CHROMIUM: ARM: tegra: DVFS resume fixes
Do not consider clocks which are not enabled and do not have DVFS
enabled on them (which, at this point, is equivalent to the clock
not being prepared) when re-calculating core voltage during resume.
Furthermore, do not update core voltage until all clocks have been
accounted for.  This prevents us from temporarily under-volting a
clock before we've recorded its frequency.

Corresponding 3.10 patch was reviewed on:
https://chromium-review.googlesource.com/182563

BUG=chrome-os-partner:31210
TEST=Run LP0 suspend/resume; check that DVFS is only enabled on the
clocks it was enabled on before suspend.

Change-Id: Ifba56403723c611a704b0020f463bd5f7f90360f
Signed-off-by: Bill Huang <bilhuang@nvidia.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214740
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217656
Tested-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>