CHROMIUM: clk: tegra: add support for Tegra132 DFLL
This patch reuses the Tegra124 DFLL driver and addes the
Tegra132 cvb and tuning data.

The opp framework is also enabled by default for generating opp
table to be used by other drivers like cpufreq.

The minimum voltage is set as 890mV for now because the thermal
function is not ready yet.

BUG=chrome-os-partner:31429
TEST=Compile and boot tested with ARCH=arm64

Change-Id: Id900cc17a3330238d663bee6726b43addc34aa59
Signed-off-by: Vince Hsu <vinceh@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/213107
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217636
Tested-by: David Riley <davidriley@chromium.org>