FROMLIST: cpufreq: Add cpufreq driver for Tegra124
Add a new cpufreq driver for Tegra124. Instead of using the PLLX as
the CPU clocksource, switch immediately to the DFLL. It allows the use
of higher clock rates, and will automatically scale the CPU voltage as
well. Besides the CPU clocksource switch, we let the cpufreq-cpu0 driver
for all the cpufreq operations.

This driver also relies on the DFLL driver to fill the OPP table for the
CPU0 device, so that the cpufreq-cpu0 driver knows what frequencies to
use.

Signed-off-by: Tuomas Tynkkynen <ttynkkynen@nvidia.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
(am from http://patchwork.ozlabs.org/patch/381807/)
Signed-off-by: Vince Hsu <vinceh@nvidia.com>

BUG=chrome-os-partner:31429
TEST=Compile and boot tested with ARCH=arm64

Change-Id: Ie28b867a09e09dd6c0b3ce9adb1c5d97f83b6c7d
Reviewed-on: https://chromium-review.googlesource.com/213103
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: Vince Hsu <vinceh@nvidia.com>
Commit-Queue: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217632
Tested-by: David Riley <davidriley@chromium.org>