CHERRY-PICK: drm/anx7808: Add sanity checks to set_property
Adds a couple of checks to set_property.

The first checks if userspace is trying to set cp to "Enabled". If
this is the case, we reject it as invalid (since the proper path to
enabled is via "Desired").

The second check returns EALREADY if userspace is trying to set the
property to "Desired" when it's already been set.

BUG=chromium:406474
TEST=Run `xrandr -d :0 --set 'Content Protection' Desired` twice,
	the property value should remain "Enabled" on the second
	run.

Change-Id: I391f635a6ea88b9a67bb0b1742cb5e150abe3d94
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/214260
Reviewed-by: Timothy Drews <tdrews@google.com>
Tested-by: Timothy Drews <tdrews@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 68ab0259ac724990c807185c4ce816538e1cf826)
Reviewed-on: https://chromium-review.googlesource.com/216192