CHROMIUM: ASoC: rt5677: use DMIC1 clock for DMIC2
The DMIC1 clock on pin D7 is present when MX-50h bit15
(Enable DMIC1 Interface) is 1.

Setting MX-50h bit14 (Enable DMIC2 Interface) to 1 enables
the DMIC2 clock which is on pin E6(GPIO5).

Since we use pin D7 for both DMIC1 and DMIC2, this patch
configures the audio routes to match.

BUG=chrome-os-partner:29649
TEST=audio recording via DMICs on samus

Change-Id: Iadb5b380c178e211a29bb58ce0aca9543458386a
Signed-off-by: Ben Zhang <benzh@chromium.org>