CHROMIUM: drm/i915: Fix enable_psr flag to actually allow PSR to be disabled
The current kernel behavior provides a flag to disable PSR but it does
not work because it is overridden based on panel PSR support.  This
changes the enable_psr flag to behave like i915_enable_rc6 where the
default (-1) is "auto" mode and it can be forced off by setting to 0.

BUG=chrome-os-partner:31833,chrome-os-partner:31836
TEST=manual testing with PSR capable panel:
1) boot without modifying i915.enable_psr (is -1) and see that PSR is
auto enabled based on the panel reporting PSR capability in DPCD.
2) boot with i915.enable_psr=0 and see that PSR is actually disabled.

Change-Id: I5f6d5b4673b85077781365e4a8f933e62030323e
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/216972
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>