CHROMIUM: HID: Created base Plantronics HID driver.
A fix for Plantronics telephony device controls interfering with
the mouse.

This version of the driver prevents Telephony pages which are not
mapped as Consumer Control applications AND are not on the Consumer Page
from being registered by the hid-input driver.

BUG=chromium:417434
TEST=Tested with multiple Plantronics devices on 'peppy' Acer C720p.

Change-Id: I7a55880d6082467274b693c48415e9331d9bbd81
Signed-off-by: JD Cole <jd.cole@plantronics.com>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/224543
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Terry Junge <terry.junge@plantronics.com>
(cherry picked from commit 7c7614b4c77c2e847454a313832df4c7e73a1877)
Reviewed-on: https://chromium-review.googlesource.com/226911