CHROMIUM: Input: elan - check if device is there before really probing
Unfortunately vendors like to interchange touchpad parts between factory
runs, but keep the same DTS, so we can't be quite sure that the device
is actually present in the system even if it is described in the DTS.

Before trying to properly initialize the touchpad iand generate bunch of
errors let's first see it there is anything at the given address. If we
get error fail silently with -ENXIO.

BUG=chrome-os-partner:34495
TEST=build and boot unit with/without Elan trackpad, verify the system
functions properly.

Change-Id: I8c26741a220a6142721436ec743a2aa7a315b163
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/236895
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/237081
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>