tpm: send TPM_SaveState at shutdown
Sending a TPM_SaveState command at shutdown ensures that the entire
session duration is considered in the logic that decrements the
Dictionary Attack counter (DA counter).  On some TPM models, the
passage of time is noted only when a commands is executed.  It is
not clear that this specific command is necessary, but it does not
hurt and this is what the manufacturer recommended.  This is expected
to reduce the occurrence of unrecoverable TPM state on the Infineon
SLB 9635 i2c (Daisy/Snow).  Infineon suggests that this may be helpful
in other TPM models as well, but I'd like to minimize changes.

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

BUG=chromium:448412
TEST=verified that code path is executed on Snow devices

Change-Id: Ib92f3e2a23e90b25a19f60a64c53d626361763a7
Reviewed-on: https://chromium-review.googlesource.com/244845
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Commit-Queue: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>