CHROMIUM: Change default value of the PSR source parameter 'idleframes' to 3
We are hearing reports of screen flickering from left to right. We are
currently conjecturing that this is PSR related and increasing idleframes
to 3 may help. Unfortunately we do not have a good repro for the issue
and therefore it will be hard to prove or disprove this.

BUG=chrome-os-partner:34320
TEST=on Samus run " cat /sys/kernel/debug/dri/0/i915_edp_psr_status"
and verify that the Performance_Counter field increments when the screen
is idle.

Original-Change-Id: I1769096f8473978f27d54dba5a33351879bed9d7
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/241171
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
(cherry picked from commit 6daa89b7c87742229b782f999692b474253f5403)

Change-Id: I321a40a63ebb92a0d7b3e7cf70eaec33378059e6
Reviewed-on: https://chromium-review.googlesource.com/241435
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Trybot-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>