FROMLIST: mac80211: only roll back station states for WDS when suspending
In normal cases (i.e. when we are fully associated), cfg80211 takes
care of removing all the stations before calling suspend in mac80211.

But in the corner case when we suspend during authentication or
association, mac80211 needs to roll back the station states.  But we
shouldn't roll back the station states in the suspend function,
because this is taken care of in other parts of the code, except for
WDS interfaces.  For AP types of interfaces, cfg80211 takes care of
disconnecting all stations before calling the driver's suspend code.
For station interfaces, this is done in the quiesce code.

For WDS interfaces we still need to do it here, so move the code into
a new switch case for WDS.

BUG=chrome-os-partner:35485
TEST=use a patch to reproduce the bug more easily, comment #25

Cc: stable@kernel.org [3.15+]
Change-Id: I2145002ee46319aa574e10a12aa5dd4f949b1427
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
(am from https://patchwork.kernel.org/patch/5680371/)
Reviewed-on: https://chromium-review.googlesource.com/242091
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277580
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Grace Kao <grace.kao@intel.com>
Tested-by: Grace Kao <grace.kao@intel.com>