CHROMIUM: pstore/ram: verify ramoops header before saving record
On some devices the persistent memory contains junk after a cold boot,
and /dev/pstore/dmesg-ramoops-* are created with random data which is
not the result of a kernel crash.

This patch adds a ramoops header check and skips any dmesg ramoops
record that does not have a valid header.

BUG=chromium:443764
TEST=platform_KernelErrorPaths and cold boot on stumpy

Change-Id: I6b5be2e934d781c44324ed183892e594988b07d1
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243390
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>