CHROMIUM: iwl7000: mac80211: start queues if driver rejected wowlan
If the driver rejects WoWLAN, restart the queues before returning
to cfg80211. cfg80211 will return to mac80211, but not before it
disconnects all interfaces. If we don't start the queues, any of
the packets needed for disconnecting won't be transmitted, which
is strange. Fix that.

BUG=N/A
TEST=make driver reject wowlan and check for deauth frame

Change-Id: I89ccfff6e22b15eaa5d781ecebe5dacf8eaba559
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/256042
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277599
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Commit-Queue: Grace Kao <grace.kao@intel.com>
Tested-by: Grace Kao <grace.kao@intel.com>